feat: Add partner services for making reservations
This commit adds two new partner services, `Partner1` and `Partner2`, for making reservations. Each partner service has a `MakeReservation` method that sends a reservation request to the respective partner API. The reservation request is constructed based on the input parameters provided. The response from the partner API is then parsed and transformed into a standardized format.

The `PartnerFactory` interface and `DefaultPartnerFactory` struct are also added to provide a way to create partner instances based on the partner ID. The `CreatePartner` method in the factory returns the appropriate partner instance based on the provided partner ID.

These changes enable the application to interact with multiple partner APIs for making reservations, allowing for more flexibility and scalability in the system.